Feat(validation): Remove validation from controllers - Created a middleware for signup validation - Refactor an article test password to meet middleware [WIP #165413101]
