+++ routes/api/auth/users.js
@@ -1,9 +1,10 @@
 import express from 'express';
 import Usercontroller from '../../../controllers/user.controller';
+import Uservalidator from '../../../helpers/userHelper';
 
 const router = express.Router();
 
-router.post('/signup', Usercontroller.addUser);
+router.post('/signup', Uservalidator.addUser, Usercontroller.addUser);
 
 router.post('/login', Usercontroller.loginUser);
